+++ CHANGES.md
@@ -172,5 +172,6 @@ Changes to nnlib2Rcpp version 0.2.7 (from 0.2.6)
 
 Changes to nnlib2Rcpp version 0.2.8 (from 0.2.7)
 - minor change in documentation of LVQ, explaining why data should be in 0 to 1 range if random values in [0 1] are used for initial weights.
+- added (due to recent supervised LVQ popularity) two LVQs wrapper functions to simplify using LVQs (functions LVQs_train, LVQs_recall). 
 
 ---